Cleanup API, bump erldpc
