+++ native/lib.rs
@@ -1,128 +1,86 @@
 use crate::bin::Bin;
+use crate::res::{DecoderRes, EncoderRes};
+use online_codes::types::CheckBlockId;
 use online_codes::{
     decode_block, new_decoder, new_decoder_with_params, new_encoder, new_encoder_with_params,
     next_block, types::StreamId,
 };
-use rustler::{Encoder, Env, NifResult, ResourceArc, Term};
-use std::sync::{RwLock, RwLockWriteGuard};
+use rustler::{error::Error, Atom, Env, NifResult, ResourceArc, Term};
 
 mod bin;
+mod res;
+
+type Drop = (u64, Bin);
 
 rustler::atoms! {
     ok,
-    error,
     incomplete,
     undefined,
 }
 
-pub struct EncoderRes(RwLock<online_codes::Encoder>);
-pub struct DecoderRes(RwLock<online_codes::decode::Decoder>);
-
-type Drop = (u64, Bin);
-
-impl EncoderRes {
-    // fn read(&self) -> RwLockReadGuard<'_, Encoder> {
-    //     self.0.read().unwrap()
-    // }
-
-    fn write(&self) -> RwLockWriteGuard<'_, online_codes::Encoder> {
-        self.0.write().unwrap()
-    }
-}
-
-impl From<online_codes::Encoder> for EncoderRes {
-    fn from(other: online_codes::Encoder) -> Self {
-        EncoderRes(RwLock::new(other))
-    }
-}
-
-impl From<online_codes::decode::Decoder> for DecoderRes {
-    fn from(other: online_codes::decode::Decoder) -> Self {
-        DecoderRes(RwLock::new(other))
-    }
-}
-
-impl DecoderRes {
-    // fn read(&self) -> RwLockReadGuard<'_, Decoder> {
-    //     self.0.read().unwrap()
-    // }
-
-    fn write(&self) -> RwLockWriteGuard<'_, online_codes::decode::Decoder> {
-        self.0.write().unwrap()
-    }
-}
-
 #[rustler::nif(name = "encoder")]
-pub fn encoder<'a>(
-    env: Env<'a>,
+pub fn encoder(
     data: Bin,
     block_size: usize,
     stream_id: StreamId,
-) -> NifResult<Term<'a>> {
+) -> NifResult<(Atom, ResourceArc<EncoderRes>)> {
     let e = new_encoder(data.0, block_size, stream_id);
-    Ok((ok(), ResourceArc::new(EncoderRes::from(e))).encode(env))
+    Ok((ok(), ResourceArc::new(EncoderRes::from(e))))
 }
 
 #[rustler::nif(name = "encoder_with_params")]
-pub fn encoder_with_params<'a>(
-    env: Env<'a>,
+pub fn encoder_with_params(
     data: Bin,
     block_size: usize,
     epsilon: f64,
     q: usize,
     stream_id: StreamId,
-) -> NifResult<Term<'a>> {
+) -> NifResult<(Atom, ResourceArc<EncoderRes>)> {
     let e = new_encoder_with_params(data.0, block_size, epsilon, q, stream_id);
-    Ok((ok(), ResourceArc::new(EncoderRes::from(e))).encode(env))
+    Ok((ok(), ResourceArc::new(EncoderRes::from(e))))
 }
 
 #[rustler::nif(name = "decoder")]
-pub fn decoder<'a>(
-    env: Env<'a>,
+pub fn decoder(
     buf_len: usize,
     block_size: usize,
     stream_id: StreamId,
-) -> NifResult<Term<'a>> {
+) -> NifResult<(Atom, ResourceArc<DecoderRes>)> {
     let d = new_decoder(buf_len, block_size, stream_id);
-    Ok((ok(), ResourceArc::new(DecoderRes::from(d))).encode(env))
+    Ok((ok(), ResourceArc::new(DecoderRes::from(d))))
 }
 
 #[rustler::nif(name = "decoder_with_params")]
-pub fn decoder_with_params<'a>(
-    env: Env<'a>,
+pub fn decoder_with_params(
     buf_len: usize,
     block_size: usize,
     epsilon: f64,
     q: usize,
     stream_id: StreamId,
-) -> NifResult<Term<'a>> {
+) -> NifResult<(Atom, ResourceArc<DecoderRes>)> {
     let d = new_decoder_with_params(buf_len, block_size, epsilon, q, stream_id);
-    Ok((ok(), ResourceArc::new(DecoderRes::from(d))).encode(env))
+    Ok((ok(), ResourceArc::new(DecoderRes::from(d))))
 }
 
 #[rustler::nif(name = "next_drop")]
-pub fn next_drop<'a>(env: Env<'a>, enc_arc: ResourceArc<EncoderRes>) -> NifResult<Term<'a>> {
+pub fn next_drop(enc_arc: ResourceArc<EncoderRes>) -> NifResult<(Atom, (CheckBlockId, Bin))> {
     let mut enc = enc_arc.write();
     match next_block(&mut enc) {
         Some(block) => {
             let (check_block_id, check_block) = block;
-            Ok((ok(), (check_block_id, Bin(check_block))).encode(env))
+            Ok((ok(), (check_block_id, Bin(check_block))))
         }
 
-        None => Ok(undefined().encode(env)),
+        None => Err(Error::Term(Box::new(undefined()))),
     }
 }
 
 #[rustler::nif(name = "decode_drop")]
-pub fn decode_drop<'a>(
-    env: Env<'a>,
-    drop: Drop,
-    dec_arc: ResourceArc<DecoderRes>,
-) -> NifResult<Term<'a>> {
+pub fn decode_drop(drop: Drop, dec_arc: ResourceArc<DecoderRes>) -> NifResult<(Atom, Bin)> {
     let mut dec = dec_arc.write();
     match decode_block((drop.0, (drop.1).0), &mut dec) {
-        Some(data) => Ok((ok(), Bin(data)).encode(env)),
-        None => Ok((error(), incomplete()).encode(env)),
+        Some(data) => Ok((ok(), Bin(data))),
+        None => Err(Error::Term(Box::new(incomplete()))),
     }
 }
