+++ src/erlang_oc.erl
@@ -19,7 +19,9 @@
 %% Encoder
 %% ==================================================================
 
--spec encoder(Data :: binary(), BlockSize :: pos_integer(), StreamID :: non_neg_integer())-> Encoder :: reference().
+-spec encoder(Data :: binary(),
+              BlockSize :: pos_integer(),
+              StreamID :: non_neg_integer()) -> {ok, Encoder :: reference()}.
 encoder(_Data, _BlockSize, _StreamID) ->
     not_loaded(?LINE).
 
@@ -28,11 +30,13 @@ encoder(_Data, _BlockSize, _StreamID) ->
         BlockSize :: pos_integer(),
         Epsilon :: float(),
         Q :: pos_integer(),
-        StreamID :: non_neg_integer())-> Encoder :: reference().
+        StreamID :: non_neg_integer()) -> {ok, Encoder :: reference()}.
 encoder_with_params(_Data, _BlockSize, _Epsilon, _Q, _StreamID) ->
     not_loaded(?LINE).
 
--spec decoder(BufLen :: non_neg_integer(), BlockSize :: pos_integer(), StreamID :: non_neg_integer())-> Decoder :: reference().
+-spec decoder(BufLen :: non_neg_integer(),
+              BlockSize :: pos_integer(),
+              StreamID :: non_neg_integer())-> {ok, Decoder :: reference()}.
 decoder(_BufLen, _BlockSize, _StreamID) ->
     not_loaded(?LINE).
 
@@ -41,15 +45,16 @@ decoder(_BufLen, _BlockSize, _StreamID) ->
         BlockSize :: pos_integer(),
         Epsilon :: float(),
         Q :: pos_integer(),
-        StreamID :: non_neg_integer())-> Decoder :: reference().
+        StreamID :: non_neg_integer()) -> {ok, Decoder :: reference()}.
 decoder_with_params(_BufLen, _BlockSize, _Epsilon, _Q, _StreamID) ->
     not_loaded(?LINE).
 
--spec next_drop(Encoder :: reference()) -> drop() | undefined.
+-spec next_drop(Encoder :: reference()) -> {ok, drop()} | {error, undefined}.
 next_drop(_Encoder) ->
     not_loaded(?LINE).
 
--spec decode_drop(Drop :: drop(), Decoder :: reference()) -> Data :: binary() | undefined.
+-spec decode_drop(Drop :: drop(), Decoder :: reference()) -> {ok, Data :: binary()} |
+                                                             {error, incomplete}.
 decode_drop(_Drop, _Decoder) ->
     not_loaded(?LINE).
